Move notifier wrappers to new file
